# docs/swagger.yaml
basePath: /
definitions:
  app.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      msg:
        type: string
    type: object
host: host
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server Petstore server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /api/v1/articles:
    get:
      parameters:
      - description: TagID
        in: body
        name: tag_id
        schema:
          $ref: '#/definitions/int'
          type: object
      - description: State
        in: body
        name: state
        schema:
          $ref: '#/definitions/int'
          type: object
      - description: CreatedBy
        in: body
        name: created_by
        schema:
          $ref: '#/definitions/int'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
            type: object
      summary: Get multiple articles
    post:
      parameters:
      - description: TagID
        in: body
        name: tag_id
        required: true
        schema:
          $ref: '#/definitions/int'
          type: object
      - description: Title
        in: body
        name: title
        required: true
        schema:
          $ref: '#/definitions/string'
          type: object
      - description: Desc
        in: body
        name: desc
        required: true
        schema:
          $ref: '#/definitions/string'
          type: object
      - description: Content
        in: body
        name: content
        required: true
        schema:
          $ref: '#/definitions/string'
          type: object
      - description: CreatedBy
        in: body
        name: created_by
        required: true
        schema:
          $ref: '#/definitions/string'
          type: object
      - description: State
        in: body
        name: state
        required: true
        schema:
          $ref: '#/definitions/int'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
            type: object
      summary: Add article
  /api/v1/articles/{id}:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
            type: object
      summary: Delete article
    get:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
            type: object
      summary: Get a single article
    put:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: TagID
        in: body
        name: tag_id
        schema:
          $ref: '#/definitions/string'
          type: object
      - description: Title
        in: body
        name: title
        schema:
          $ref: '#/definitions/string'
          type: object
      - description: Desc
        in: body
        name: desc
        schema:
          $ref: '#/definitions/string'
          type: object
      - description: Content
        in: body
        name: content
        schema:
          $ref: '#/definitions/string'
          type: object
      - description: ModifiedBy
        in: body
        name: modified_by
        required: true
        schema:
          $ref: '#/definitions/string'
          type: object
      - description: State
        in: body
        name: state
        schema:
          $ref: '#/definitions/int'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
            type: object
      summary: Update article
  /api/v1/tags:
    get:
      parameters:
      - description: Name
        in: query
        name: name
        type: string
      - description: State
        in: query
        name: state
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
            type: object
      summary: Get multiple article tags
    post:
      parameters:
      - description: Name
        in: body
        name: name
        required: true
        schema:
          $ref: '#/definitions/string'
          type: object
      - description: State
        in: body
        name: state
        schema:
          $ref: '#/definitions/int'
          type: object
      - description: CreatedBy
        in: body
        name: created_by
        schema:
          $ref: '#/definitions/int'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
            type: object
      summary: Add article tag
  /api/v1/tags/{id}:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
            type: object
      summary: Delete article tag
    put:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: Name
        in: body
        name: name
        required: true
        schema:
          $ref: '#/definitions/string'
          type: object
      - description: State
        in: body
        name: state
        schema:
          $ref: '#/definitions/int'
          type: object
      - description: ModifiedBy
        in: body
        name: modified_by
        required: true
        schema:
          $ref: '#/definitions/string'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
            type: object
      summary: Update tag
  /auth:
    get:
      parameters:
      - description: username
        in: query
        name: username
        required: true
        type: string
      - description: password
        in: query
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
            type: object
      summary: Get Auth
  /upload:
    post:
      parameters:
      - description: Image File
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
            type: object
      summary: Upload Image
swagger: "2.0"